moved poll results files to result packge and created a new package for create poll
