+++ .gitignore
@@ -124,7 +124,7 @@ proguard/
 # Keystore files
 
 # Google Services (e.g. APIs or Firebase)
-# google-services.json
+google-services.json
 
 # Android Patch
