+++ likeminds-feed-android-core/src/main/java/com/likeminds/feed/android/core/poll/result/model/LMFeedPollViewData.kt
@@ -1,4 +1,4 @@
-package com.likeminds.feed.android.core.poll.model
+package com.likeminds.feed.android.core.poll.result.model
 
 import android.content.Context
 import android.os.Parcelable
