+++ likeminds-feed-android-core/src/main/java/com/likeminds/feed/android/core/ui/widgets/poll/adapter/databinder/LMFeedItemPollOptionsViewDataBinder.kt
@@ -3,7 +3,7 @@ package com.likeminds.feed.android.core.ui.widgets.poll.adapter.databinder
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import com.likeminds.feed.android.core.databinding.LmFeedItemPollOptionsBinding
-import com.likeminds.feed.android.core.poll.model.LMFeedPollOptionViewData
+import com.likeminds.feed.android.core.poll.result.model.LMFeedPollOptionViewData
 import com.likeminds.feed.android.core.ui.widgets.poll.adapter.LMFeedPollOptionsAdapterListener
 import com.likeminds.feed.android.core.ui.widgets.poll.view.LMFeedPollOptionView
 import com.likeminds.feed.android.core.utils.LMFeedStyleTransformer
