+++ likeminds-feed-android-core/src/main/java/com/likeminds/feed/android/core/utils/LMFeedStyleTransformer.kt
@@ -4,7 +4,7 @@ import com.likeminds.feed.android.core.activityfeed.style.LMFeedActivityFeedFrag
 import com.likeminds.feed.android.core.delete.style.LMFeedAdminDeleteDialogFragmentStyle
 import com.likeminds.feed.android.core.delete.style.LMFeedSelfDeleteDialogFragmentStyle
 import com.likeminds.feed.android.core.likes.style.LMFeedLikesFragmentViewStyle
-import com.likeminds.feed.android.core.poll.style.LMFeedPollResultsFragmentViewStyle
+import com.likeminds.feed.android.core.poll.result.style.LMFeedPollResultsFragmentViewStyle
 import com.likeminds.feed.android.core.post.create.viewstyle.LMFeedCreatePostFragmentViewStyle
 import com.likeminds.feed.android.core.post.detail.style.LMFeedPostDetailFragmentViewStyle
 import com.likeminds.feed.android.core.post.edit.style.LMFeedEditPostFragmentViewStyle
